# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Solunex Lab Backend",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/results/instantiate":  {
                                                   "post":  {
                                                                "tags":  [
                                                                             "results"
                                                                         ],
                                                                "summary":  "Instantiate Result",
                                                                "operationId":  "instantiate_result_api_results_instantiate_post",
                                                                "requestBody":  {
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/ResultInstantiate"
                                                                                                                                        }
                                                                                                                         }
                                                                                                },
                                                                                    "required":  true
                                                                                },
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "Successful Response",
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/TestResultOut"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          }
                                                                                          },
                                                                                  "422":  {
                                                                                              "description":  "Validation Error",
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                               },
                  "/api/results/{result_id}":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "results"
                                                                        ],
                                                               "summary":  "Get Result",
                                                               "operationId":  "get_result_api_results__result_id__get",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "result_id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "title":  "Result Id"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Successful Response",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/TestResultOut"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "422":  {
                                                                                             "description":  "Validation Error",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         }
                                                                             }
                                                           }
                                               },
                  "/api/results/{result_id}/values":  {
                                                          "patch":  {
                                                                        "tags":  [
                                                                                     "results"
                                                                                 ],
                                                                        "summary":  "Update Result Values",
                                                                        "operationId":  "update_result_values_api_results__result_id__values_patch",
                                                                        "parameters":  [
                                                                                           {
                                                                                               "name":  "result_id",
                                                                                               "in":  "path",
                                                                                               "required":  true,
                                                                                               "schema":  {
                                                                                                              "type":  "integer",
                                                                                                              "title":  "Result Id"
                                                                                                          }
                                                                                           }
                                                                                       ],
                                                                        "requestBody":  {
                                                                                            "required":  true,
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/ResultUpdateValues"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        },
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "Successful Response",
                                                                                                      "content":  {
                                                                                                                      "application/json":  {
                                                                                                                                               "schema":  {
                                                                                                                                                              "$ref":  "#/components/schemas/TestResultOut"
                                                                                                                                                          }
                                                                                                                                           }
                                                                                                                  }
                                                                                                  },
                                                                                          "422":  {
                                                                                                      "description":  "Validation Error",
                                                                                                      "content":  {
                                                                                                                      "application/json":  {
                                                                                                                                               "schema":  {
                                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                          }
                                                                                                                                           }
                                                                                                                  }
                                                                                                  }
                                                                                      }
                                                                    }
                                                      },
                  "/api/results/{result_id}/status":  {
                                                          "patch":  {
                                                                        "tags":  [
                                                                                     "results"
                                                                                 ],
                                                                        "summary":  "Set Result Status",
                                                                        "operationId":  "set_result_status_api_results__result_id__status_patch",
                                                                        "parameters":  [
                                                                                           {
                                                                                               "name":  "result_id",
                                                                                               "in":  "path",
                                                                                               "required":  true,
                                                                                               "schema":  {
                                                                                                              "type":  "integer",
                                                                                                              "title":  "Result Id"
                                                                                                          }
                                                                                           },
                                                                                           {
                                                                                               "name":  "x-role",
                                                                                               "in":  "header",
                                                                                               "required":  false,
                                                                                               "schema":  {
                                                                                                              "type":  "string",
                                                                                                              "default":  "labtech",
                                                                                                              "title":  "X-Role"
                                                                                                          }
                                                                                           }
                                                                                       ],
                                                                        "requestBody":  {
                                                                                            "required":  true,
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/ResultSetStatus"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        },
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "Successful Response",
                                                                                                      "content":  {
                                                                                                                      "application/json":  {
                                                                                                                                               "schema":  {
                                                                                                                                                              "$ref":  "#/components/schemas/TestResultOut"
                                                                                                                                                          }
                                                                                                                                           }
                                                                                                                  }
                                                                                                  },
                                                                                          "422":  {
                                                                                                      "description":  "Validation Error",
                                                                                                      "content":  {
                                                                                                                      "application/json":  {
                                                                                                                                               "schema":  {
                                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                          }
                                                                                                                                           }
                                                                                                                  }
                                                                                                  }
                                                                                      }
                                                                    }
                                                      },
...
                  "/portal/results":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "portal"
                                                               ],
                                                      "summary":  "Portal List Results",
                                                      "operationId":  "portal_list_results_portal_results_get",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "authorization",
                                                                             "in":  "header",
                                                                             "required":  false,
                                                                             "schema":  {
                                                                                            "anyOf":  [
                                                                                                          {
                                                                                                              "type":  "string"
                                                                                                          },
                                                                                                          {
                                                                                                              "type":  "null"
                                                                                                          }
                                                                                                      ],
                                                                                            "title":  "Authorization"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Successful Response",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "array",
                                                                                                                                            "items":  {
                                                                                                                                                          "$ref":  "#/components/schemas/PortalResultItem"
                                                                                                                                                      },
                                                                                                                                            "title":  "Response Portal List Results Portal Results Get"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                        "422":  {
                                                                                    "description":  "Validation Error",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                }
                                                                    }
                                                  }
                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ResultInstantiate":  {
                                                                 "properties":  {
                                                                                    "patient_id":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "Patient Id"
                                                                                                   },
                                                                                    "template_id":  {
                                                                                                        "type":  "integer",
                                                                                                        "title":  "Template Id"
                                                                                                    }
                                                                                },
                                                                 "type":  "object",
                                                                 "required":  [
                                                                                  "patient_id",
                                                                                  "template_id"
                                                                              ],
                                                                 "title":  "ResultInstantiate"
                                                             },
...
                                   }
                   }
}